\name{gd_rates}
\docType{data}
\alias{gd_rates}
\title{Growth and decay rates}
\description{
 A data.frame containing growth and decay rates for water use drivers by sector and region. 
}
\usage{drivers}
\format{Data.frame with 4 columns (region, sector, g=growth rate, d=decay rate)}
\source{See original script \code{data-raw/RPA-Water_MEllis.R}}
\section{Data}{\verb{
> gd_rates
  region sector       g      d
1   East     dp -0.0066 -0.030
2   West     dp -0.0035 -0.030
3   East     ic -0.0331 -0.035
4   West     ic -0.0578 -0.042
}  }

\keyword{datasets}