\name{wd2010}
\docType{data}
\alias{wd2010}
\title{Driver initial values}
\description{
 A data.frame containing values from 2010 for water use drivers for each sector
}
\usage{drivers}
\format{Data.frame with 4 columns (WRR, ASR, sector, wd=water demand)}
\source{See original script \code{data-raw/RPA-Water_MEllis.R}}
\section{Data}{\verb{
> wd2010
   WRR  ASR    sector   wd
1   14 1401        dp   32
2   14 1402        dp   67
3   14 1403        dp   52
4   15 1501        dp   52
5   15 1502        dp  459
6   15 1503        dp 1043
7   14 1401        ic   16
8   14 1402        ic   36
9   14 1403        ic   21
10  15 1501        ic   15
11  15 1502        ic  188
12  15 1503        ic  412
13  14 1401    thermo  101
14  14 1402    thermo    7
15  14 1403    thermo   62
16  15 1501    thermo   44
17  15 1502    thermo   42
18  15 1503    thermo   31
19  14 1401     irrig 3161
20  14 1402     irrig 2148
21  14 1403     irrig 1138
22  15 1501     irrig   72
23  15 1502     irrig 1603
24  15 1503     irrig 3364
25  14 1401 livestock   46
26  14 1402 livestock   37
27  14 1403 livestock   40
28  15 1501 livestock    9
29  15 1502 livestock    5
30  15 1503 livestock   19
}  }

\keyword{datasets}