\name{gd_rates}
\docType{data}
\alias{gd_rates}
\title{Growth and decay rates}
\description{
 A data.frame containing growth and decay rates for water use drivers by sector and region. 
}
\usage{drivers}
\format{Data.frame with 4 columns (region, sector, g=growth rate, d=decay rate)}
\source{See original script \code{data-raw/RPA-Water_MEllis.R}}

\keyword{datasets}